Renamed classes, made the dialog more generic.
