+++ GoogleCloudExtension/GoogleCloudExtension/CloudExplorerSources/Gce/GceInstanceViewModel.cs
@@ -22,6 +22,7 @@ using GoogleCloudExtension.ManageWindowsCredentials;
 using GoogleCloudExtension.OAuth;
 using GoogleCloudExtension.TerminalServer;
 using GoogleCloudExtension.Utils;
+using GoogleCloudExtension.WindowsCredentialsChooser;
 using Microsoft.Win32;
 using System;
 using System.Collections.Generic;
@@ -330,7 +331,7 @@ namespace GoogleCloudExtension.CloudExplorerSources.Gce
         {
             ExtensionAnalytics.ReportCommand(CommandName.OpenTerminalServerSessionForGceInstanceCommand, CommandInvocationSource.Button);
 
-            TerminalServerManagerWindow.PromptUser(_instance);
+            var credentials = WindowsCredentialsChooserWindow.PromptUser(_instance);
         }
 
         private void OnOpenWebsite()
