+++ GoogleCloudExtension/GoogleCloudExtension/WindowsCredentialsChooser/WindowsCredentialsChooserWindowContent.xaml
@@ -1,14 +1,14 @@
-﻿<UserControl x:Class="GoogleCloudExtension.TerminalServer.TerminalServerManagerWindowContent"
+﻿<UserControl x:Class="GoogleCloudExtension.WindowsCredentialsChooser.WindowsCredentialsChooserWindowContent"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:local="clr-namespace:GoogleCloudExtension.TerminalServer"
+             xmlns:local="clr-namespace:GoogleCloudExtension.WindowsCredentialsChooser"
              xmlns:theming="clr-namespace:GoogleCloudExtension.Theming"
              xmlns:ext="clr-namespace:GoogleCloudExtension"
              mc:Ignorable="d" 
              d:DesignHeight="300" d:DesignWidth="300"
-             d:DataContext="{d:DesignInstance {x:Type local:TerminalServerManagerViewModel}}">
+             d:DataContext="{d:DesignInstance {x:Type local:WindowsCredentialsChooserViewModel}}">
 
     <UserControl.Resources>
         <StackPanel x:Key="content">
